Crate finitediff

This crate contains a wide range of methods for the calculation of gradients, Jacobians and Hessians using forward and central differences. The methods have been implemented for input vectors of the type Vec<f64> and ndarray::Array1<f64>. Central differences are more accurate but require more evaluations of the cost function and are therefore computationally more expensive.

References

Jorge Nocedal and Stephen J. Wright (2006). Numerical Optimization. Springer. ISBN 0-387-30303-0.

Usage

Add this to your dependencies section of Cargo.toml:

[dependencies]
finitediff = "0.1.4"

To use the FiniteDiff trait implementations on the ndarray types, please activate the ndarray feature:

[dependencies]
finitediff = { version = "0.1.4", features = ["ndarray"] }

Examples

Calculation of the gradient

For Vec<f64>

For ndarray::Array1<f64>

Calculation of the Jacobian

Full Jacobian

Product of the Jacobian J(x) with a vector p

Sparse Jacobian

Calculation of the Hessian

Full Hessian

Product of the Hessian H(x) with a vector p

Calculation of the Hessian without knowledge of the gradient

Calculation of the sparse Hessian without knowledge of the gradient

Calculation of the gradient

This section illustrates the computation of a gradient of a function f at a point x of type Vec<f64>. Using forward differences requires n+1 evaluations of the function f while central differences requires 2*n evaluations.

For Vec<f64>

use finitediff::vec;

// Define function `f(x)`
let f = |x: &Vec<f64>| -> Result<f64, anyhow::Error> {
    // ...
};

// Point at which gradient should be calculated
let x = vec![1.0f64, 1.0];

// Calculate gradient of `f` at `x` using forward differences
let g_forward = vec::forward_diff(&f);
let grad_forward = g_forward(&x)?;

// Calculate gradient of `f` at `x` using central differences
let g_central = vec::central_diff(&f);
let grad_central = g_central(&x)?;

For ndarray::Array1<f64>

use ndarray::{array, Array1};
use finitediff::ndarr;

// Define cost function `f(x)`
let f = |x: &Array1<f64>| -> Result<f64, anyhow::Error> {
    // ...
};

// Point at which gradient should be calculated
let x = array![1.0f64, 1.0];

// Calculate gradient of `f` at `x` using forward differences
let g_forward = ndarr::forward_diff(&f);
let grad_forward = g_forward(&x)?;

// Calculate gradient of `f` at `x` using central differences
let g_central = ndarr::central_diff(&f);
let grad_central = g_central(&x)?;

Calculation of the Jacobian

Note that the same interface is also implemented for ndarray::Array1<f64> (not shown).

Full Jacobian

The calculation of the full Jacobian requires n+1 evaluations of the function f.

use finitediff::vec;

let f = |x: &Vec<f64>| -> Result<Vec<f64>, anyhow::Error> {
    // ...
};

let x = vec![1.0f64, 1.0, 1.0, 1.0, 1.0, 1.0];

// Using forward differences
let j_forward = vec::forward_jacobian(&f);
let jacobian_forward = j_forward(&x)?;

// Using central differences
let j_central = vec::central_jacobian(&f);
let jacobian_central = j_central(&x)?;

Product of the Jacobian J(x) with a vector p

Directly computing J(x)*p can be much more efficient than computing J(x) first and then multiplying it with p. While computing the full Jacobian J(x) requires n+1 evaluations of f, J(x)*p only requires 2.

use finitediff::vec;

let f = |x: &Vec<f64>| -> Result<Vec<f64>, anyhow::Error> {
    // ...
};

let x = vec![1.0f64, 1.0, 1.0, 1.0, 1.0, 1.0];
let p = vec![1.0f64, 2.0, 3.0, 4.0, 5.0, 6.0];

// using forward differences
let j_forward = vec::forward_jacobian_vec_prod(&f);
let jacobian_forward = j_forward(&x, &p)?;

// using central differences
let j_central = vec::central_jacobian_vec_prod(&f);
let jacobian_central = j_central(&x, &p)?;

Sparse Jacobian

If the Jacobian is sparse its structure can be exploited using perturbation vectors. See Nocedal & Wright for details.

use finitediff::{vec, PerturbationVector};

let f = |x: &Vec<f64>| -> Result<Vec<f64>, anyhow::Error> {
    // ...
};

let x = vec![1.0f64, 1.0, 1.0, 1.0, 1.0, 1.0];

let pert = vec![
    PerturbationVector::new()
        .add(0, vec![0, 1])
        .add(3, vec![2, 3, 4]),
    PerturbationVector::new()
        .add(1, vec![0, 1, 2])
        .add(4, vec![3, 4, 5]),
    PerturbationVector::new()
        .add(2, vec![1, 2, 3])
        .add(5, vec![4, 5]),
];

// using forward differences
let j_forward = vec::forward_jacobian_pert(&f);
let jacobian_forward = j_forward(&x, &pert)?;

// using central differences
let j_central = vec::central_jacobian_pert(&f);
let jacobian_central = j_central(&x, &pert)?;

Calculation of the Hessian

Note that the same interface is also implemented for ndarray::Array1<f64> (not shown).

Full Hessian

use finitediff::vec;

let g = |x: &Vec<f64>| -> Result<Vec<f64>, anyhow::Error> {
    // ...
};

let x = vec![1.0f64, 1.0, 1.0, 1.0];

// using forward differences
let h_forward = vec::forward_hessian(&g);
let hessian_forward = h_forward(&x)?;

// using central differences
let h_central = vec::central_hessian(&g);
let hessian_central = h_central(&x)?;

Product of the Hessian H(x) with a vector p

use finitediff::vec;

let g = |x: &Vec<f64>| -> Result<Vec<f64>, anyhow::Error> {
    // ...
};

let x = vec![1.0f64, 1.0, 1.0, 1.0];
let p = vec![2.0, 3.0, 4.0, 5.0];

// using forward differences
let h_forward = vec::forward_hessian_vec_prod(&g);
let hessian_forward = h_forward(&x, &p)?;

// using forward differences
let h_central = vec::central_hessian_vec_prod(&g);
let hessian_central = h_central(&x, &p)?;

Calculation of the Hessian without knowledge of the gradient

use finitediff::vec;

let f = |x: &Vec<f64>| -> Result<f64, anyhow::Error> {
    // ...
};

let x = vec![1.0f64, 1.0, 1.0, 1.0];

let h = vec::forward_hessian_nograd(&f);
let hessian = h(&x)?;

Calculation of the sparse Hessian without knowledge of the gradient

use finitediff::vec;

let f = |x: &Vec<f64>| -> Result<f64, anyhow::Error> {
    // ...
};

let x = vec![1.0f64, 1.0, 1.0, 1.0];

// Indices at which the Hessian should be evaluated. All other
// elements of the Hessian will be zero
let indices = vec![[1, 1], [2, 3], [3, 3]];

let h = vec::forward_hessian_nograd_sparse(&f);
let hessian = h(&x, indices)?;

Modules

array

vec

Structs

PerturbationVector

Perturbation Vector for the accelerated computation of the Jacobian.

Type Aliases

PerturbationVectors

A collection of PerturbationVectors