Crate finitediff

Finite Differentiation

References

Jorge Nocedal and Stephen J. Wright (2006). Numerical Optimization. Springer. ISBN 0-387-30303-0.

Examples

Calculation of the gradient

Using forward differences

use finitediff::FiniteDiff;

// Define cost function `f(x) = x_0 + x_1^2`
let f = |x: &Vec<f64>| x[0] + x[1].powi(2);

// Point at which gradient should be calculated
let x = vec![1.0f64, 1.0];

// Calculate gradient of `f` at `x` using forward differences
let grad = x.forward_diff(&f);

// Desired solution
let res = vec![1.0f64, 2.0];

// Check result
(0..2)
    .map(|i| assert!((res[i] - grad[i]).abs() < 1e-6))
    .count();

Using central differences

use finitediff::FiniteDiff;

// Define cost function `f(x) = x_0 + x_1^2`
let f = |x: &Vec<f64>| x[0] + x[1].powi(2);

// Point at which gradient should be calculated
let x = vec![1.0f64, 1.0];

// Calculate gradient of `f` at `x` using central differences
let grad = x.central_diff(&f);

// Desired solution
let res = vec![1.0f64, 2.0];

// Check result
(0..2)
    .map(|i| assert!((res[i] - grad[i]).abs() < 1e-6))
    .count();

Calculation of the Jacobian

Full Jacobian

use finitediff::FiniteDiff;

let f = |x: &Vec<f64>| {
    vec![
        2.0 * (x[1].powi(3) - x[0].powi(2)),
        3.0 * (x[1].powi(3) - x[0].powi(2)) + 2.0 * (x[2].powi(3) - x[1].powi(2)),
        3.0 * (x[2].powi(3) - x[1].powi(2)) + 2.0 * (x[3].powi(3) - x[2].powi(2)),
        3.0 * (x[3].powi(3) - x[2].powi(2)) + 2.0 * (x[4].powi(3) - x[3].powi(2)),
        3.0 * (x[4].powi(3) - x[3].powi(2)) + 2.0 * (x[5].powi(3) - x[4].powi(2)),
        3.0 * (x[5].powi(3) - x[4].powi(2)),
    ]
};

let x = vec![1.0f64, 1.0, 1.0, 1.0, 1.0, 1.0];

let jacobian = x.forward_jacobian(&f);

let res = vec![
    vec![-4.0, -6.0, 0.0, 0.0, 0.0, 0.0],
    vec![6.0, 5.0, -6.0, 0.0, 0.0, 0.0],
    vec![0.0, 6.0, 5.0, -6.0, 0.0, 0.0],
    vec![0.0, 0.0, 6.0, 5.0, -6.0, 0.0],
    vec![0.0, 0.0, 0.0, 6.0, 5.0, -6.0],
    vec![0.0, 0.0, 0.0, 0.0, 6.0, 9.0],
];

(0..6)
    .zip(0..6)
    .map(|(i, j)| assert!((res[i][j] - jacobian[i][j]).abs() < 1e-6))
    .count();

Product of the Jacobian J(x) with a vector p

use finitediff::FiniteDiff;

let f = |x: &Vec<f64>| {
    vec![
        2.0 * (x[1].powi(3) - x[0].powi(2)),
        3.0 * (x[1].powi(3) - x[0].powi(2)) + 2.0 * (x[2].powi(3) - x[1].powi(2)),
        3.0 * (x[2].powi(3) - x[1].powi(2)) + 2.0 * (x[3].powi(3) - x[2].powi(2)),
        3.0 * (x[3].powi(3) - x[2].powi(2)) + 2.0 * (x[4].powi(3) - x[3].powi(2)),
        3.0 * (x[4].powi(3) - x[3].powi(2)) + 2.0 * (x[5].powi(3) - x[4].powi(2)),
        3.0 * (x[5].powi(3) - x[4].powi(2)),
    ]
};

let x = vec![1.0f64, 1.0, 1.0, 1.0, 1.0, 1.0];
let p = vec![1.0f64, 2.0, 3.0, 4.0, 5.0, 6.0];

let jacobian = x.forward_jacobian_vec_prod(&f, &p);

let res = vec![8.0, 22.0, 27.0, 32.0, 37.0, 24.0];

(0..6)
    .map(|i| assert!((res[i] - jacobian[i]).abs() < 1e-4))
    .count();

Sparse Jacobian

use finitediff::{FiniteDiff, PerturbationVector};

let f = |x: &Vec<f64>| {
    vec![
        2.0 * (x[1].powi(3) - x[0].powi(2)),
        3.0 * (x[1].powi(3) - x[0].powi(2)) + 2.0 * (x[2].powi(3) - x[1].powi(2)),
        3.0 * (x[2].powi(3) - x[1].powi(2)) + 2.0 * (x[3].powi(3) - x[2].powi(2)),
        3.0 * (x[3].powi(3) - x[2].powi(2)) + 2.0 * (x[4].powi(3) - x[3].powi(2)),
        3.0 * (x[4].powi(3) - x[3].powi(2)) + 2.0 * (x[5].powi(3) - x[4].powi(2)),
        3.0 * (x[5].powi(3) - x[4].powi(2)),
    ]
};

let x = vec![1.0f64, 1.0, 1.0, 1.0, 1.0, 1.0];

let pert = vec![
    PerturbationVector::new()
        .add(0, vec![0, 1])
        .add(3, vec![2, 3, 4]),
    PerturbationVector::new()
        .add(1, vec![0, 1, 2])
        .add(4, vec![3, 4, 5]),
    PerturbationVector::new()
        .add(2, vec![1, 2, 3])
        .add(5, vec![4, 5]),
];

let jacobian = x.forward_jacobian_pert(&f, pert);

let res = vec![
    vec![-4.0, -6.0, 0.0, 0.0, 0.0, 0.0],
    vec![6.0, 5.0, -6.0, 0.0, 0.0, 0.0],
    vec![0.0, 6.0, 5.0, -6.0, 0.0, 0.0],
    vec![0.0, 0.0, 6.0, 5.0, -6.0, 0.0],
    vec![0.0, 0.0, 0.0, 6.0, 5.0, -6.0],
    vec![0.0, 0.0, 0.0, 0.0, 6.0, 9.0],
];

(0..6)
    .zip(0..6)
    .map(|(i, j)| assert!((res[i][j] - jacobian[i][j]).abs() < 1e-6))
    .count();

Calculation of the Hessian

Full Hessian

use finitediff::FiniteDiff;

let f = |x: &Vec<f64>| x[0] + x[1].powi(2) + x[2] * x[3].powi(2);

let x = vec![1.0f64, 1.0, 1.0, 1.0];

let hessian = x.forward_hessian(&|d| d.forward_diff(&f));

let res = vec![
    vec![0.0, 0.0, 0.0, 0.0],
    vec![0.0, 2.0, 0.0, 0.0],
    vec![0.0, 0.0, 0.0, 2.0],
    vec![0.0, 0.0, 2.0, 2.0],
];

(0..4)
    .zip(0..4)
    .map(|(i, j)| assert!((res[i][j] - hessian[i][j]).abs() < 1e-6))
    .count();

Product of Hessian H(x) with a vector p

use finitediff::FiniteDiff;

let f = |x: &Vec<f64>| x[0] + x[1].powi(2) + x[2] * x[3].powi(2);

let x = vec![1.0f64, 1.0, 1.0, 1.0];
let p = vec![2.0, 3.0, 4.0, 5.0];

let hessian = x.forward_hessian_vec_prod(&|d| d.forward_diff(&f), &p);

let res = vec![0.0, 6.0, 10.0, 18.0];

(0..4)
    .map(|i| assert!((res[i] - hessian[i]).abs() < 1e-6))
    .count();

Calculation of the Hessian without knowledge of the gradient

use finitediff::FiniteDiff;

let f = |x: &Vec<f64>| x[0] + x[1].powi(2) + x[2] * x[3].powi(2);

let x = vec![1.0f64, 1.0, 1.0, 1.0];

let hessian = x.forward_hessian_nograd(&f);

let res = vec![
    vec![0.0, 0.0, 0.0, 0.0],
    vec![0.0, 2.0, 0.0, 0.0],
    vec![0.0, 0.0, 0.0, 2.0],
    vec![0.0, 0.0, 2.0, 2.0],
];

(0..4)
    .zip(0..4)
    .map(|(i, j)| assert!((res[i][j] - hessian[i][j]).abs() < 1e-6))
    .count();

Calculation of the sparse Hessian without knowledge of the gradient

use finitediff::FiniteDiff;

let f = |x: &Vec<f64>| x[0] + x[1].powi(2) + x[2] * x[3].powi(2);

let x = vec![1.0f64, 1.0, 1.0, 1.0];

let indices = vec![(1, 1), (2, 3), (3, 3)];

let hessian = x.forward_hessian_nograd_sparse(&f, indices);

let res = vec![
    vec![0.0, 0.0, 0.0, 0.0],
    vec![0.0, 2.0, 0.0, 0.0],
    vec![0.0, 0.0, 0.0, 2.0],
    vec![0.0, 0.0, 2.0, 2.0],
];

(0..4)
    .zip(0..4)
    .map(|(i, j)| assert!((res[i][j] - hessian[i][j]).abs() < 1e-6))
    .count();

Structs

PerturbationVector
Perturbation Vector for the accelerated computation of the Jacobian.

Traits

FiniteDiff

Type Definitions

PerturbationVectors
A collection of PerturbationVectors